[#3] refactor: EmailFormGenerator 로직을 enum 타입으로 개선
