+++ event-consumer/event-api/src/main/kotlin/com/meloning/megaCoffee/listener/EventListener.kt
@@ -1,6 +1,6 @@
 package com.meloning.megaCoffee.listener
 
-import com.meloning.megaCoffee.clients.javaEmail.service.EmailFormGenerator
+import com.meloning.megaCoffee.clients.javaEmail.model.EmailFormType
 import com.meloning.megaCoffee.clients.javaEmail.service.EmailSender
 import org.slf4j.LoggerFactory
 import org.springframework.amqp.rabbit.annotation.Exchange
@@ -25,6 +25,7 @@ class EventListener(
     )
     fun emailListener(@Payload payload: Map<String, String>) {
         logger.info("payload=$payload")
-        emailSender.send(EmailFormGenerator.generate(mailProperties.username, payload))
+        val emailFormType = EmailFormType.valueOf(payload["type"]!!.uppercase())
+        emailSender.send(emailFormType.getEmailFormDto(mailProperties.username, payload))
     }
 }
